fixed issue where owner can leave public event

+++ routes/events_routes/leave_public.js
@@ -5,7 +5,28 @@ leave_public.delete("/", async (req, res) => {
     console.log("/api/events/public/join")
 
     const body = req.body
-    // Leave event in DB
+    // 1. check if is owner; if yes, then can't leave, must delete
+    const find_event_query = `SELECT * FROM events_public WHERE id=${eventId}`
+    let event
+    try {
+        event = await query(find_event_query)
+    } catch (error) {
+        res.send({
+            error,
+            message: "check if is owner; if yes, then can't leave, must delete",
+            query: find_event_query
+        }).status(400)
+    }
+    event = event[0]
+
+    if (event.ownerId === body.userId) {
+        res.send({
+            message: "Error: owner attempting to leave own event. Should delete instead."
+        }).status(404)
+        return
+    }
+
+    // 2. Delete entry in table 'users_to_public' to leave event
     const leave_event_query = `DELETE FROM users_to_public WHERE userId=${body.userId} AND eventId=${body.eventId}`
     let leave_event_response
     try {
@@ -13,7 +34,7 @@ leave_public.delete("/", async (req, res) => {
     } catch (error) {
         res.send({
             error,
-            message: "Leave event in DB",
+            message: "Delete entry in table 'users_to_public' to leave event",
             query: leave_event_query
         }).status(400)
     }
